bugfix: do not emit conts for cases multiple times
This can, in rare cases, break scoping in a subtle way:
Because emitted AST nodes are cached, nodes that rely on threaded
CPS calls end up referring to unbound params when re-used in a
different context/continuation. The fix is to also cache the
MatchCase::emit() continuation and only create it once.