Accept ProjExpr in filters.
This is required to extract members of structs in a filter.
As an absolute baseline, the lhs is checked to be a legal element in a
filter, but there are potential problems, for instance with mutable
pointers, that might break in the future. If something like this comes
up we will fix it.