Support declaring external types.
Syntax (for now): `type_ext <name> =  { <arg>, ... };` where the
arguments are strings that are passed through to the thorin backend.

The implementation of this feature depends on the backend. For now, only
C supports this feature, and only a single argument is supported.
`type_ext bar = { "foo" };` will cause thorin to emit `typedef foo bar;`.
The type (`bar`) is opaque to artic and thorin. This feature is mainly
intended to interact with external APIs, not to extend the artic
programming language as a whole.